# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ForgePulse Manufacturing Execution System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /inventory/items:
    get:
      operationId: getInventoryItems
      tags: [inventory]
      summary: List inventory items
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: warehouse
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Inventory items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItemList"
    post:
      operationId: createInventoryItem
      tags: [inventory]
      summary: Create inventory item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
  /inventory/items/{id}:
    get:
      operationId: getInventoryItem
      tags: [inventory]
      summary: Get inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Inventory item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
    patch:
      operationId: updateInventoryItem
      tags: [inventory]
      summary: Update inventory item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InventoryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
...
  /procurement/suppliers:
    get:
      operationId: getSuppliers
      tags: [procurement]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Suppliers list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [procurement]
      summary: Create supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SupplierInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /procurement/purchase-orders:
    get:
      operationId: getPurchaseOrders
      tags: [procurement]
      summary: List purchase orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: supplierId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Purchase orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PurchaseOrder"
    post:
      operationId: createPurchaseOrder
      tags: [procurement]
      summary: Create purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /procurement/purchase-orders/{id}:
    patch:
      operationId: updatePurchaseOrder
      tags: [procurement]
      summary: Update purchase order status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseOrderUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
...
components:
  schemas:
...
    InventoryItem:
      type: object
      required: [id, itemCode, description, category, warehouse, uom, onHandQty, reservedQty, availableQty, reorderPoint, safetyStock, unitCost, inventoryValue, status]
      properties:
        id:
          type: integer
        itemCode:
          type: string
        description:
          type: string
        category:
          type: string
        warehouse:
          type: string
        uom:
          type: string
        onHandQty:
          type: number
        reservedQty:
          type: number
        availableQty:
          type: number
        reorderPoint:
          type: number
        safetyStock:
          type: number
        unitCost:
          type: number
        inventoryValue:
          type: number
        status:
          type: string
        abcClass:
          type: ["string", "null"]
    InventoryItemList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InventoryItem"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    InventoryItemInput:
      type: object
      required: [itemCode, description, category, warehouse, uom, onHandQty, reorderPoint, safetyStock, unitCost]
      properties:
        itemCode:
          type: string
        description:
          type: string
        category:
          type: string
        warehouse:
          type: string
        uom:
          type: string
        onHandQty:
          type: number
        reorderPoint:
          type: number
        safetyStock:
          type: number
        unitCost:
          type: number
    InventoryItemUpdate:
      type: object
      properties:
        description:
          type: string
        category:
          type: string
        warehouse:
          type: string
        onHandQty:
          type: number
        reorderPoint:
          type: number
        safetyStock:
          type: number
        unitCost:
          type: number
        status:
          type: string
...
    Supplier:
      type: object
      required: [id, supplierCode, name, category, country, contactName, contactEmail, status, rating]
      properties:
        id:
          type: integer
        supplierCode:
          type: string
        name:
          type: string
        category:
          type: string
        country:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: ["string", "null"]
        status:
          type: string
        rating:
          type: number
        totalOrders:
          type: integer
        totalValue:
          type: number
    SupplierInput:
      type: object
      required: [supplierCode, name, category, country, contactName, contactEmail]
      properties:
        supplierCode:
          type: string
        name:
          type: string
        category:
          type: string
        country:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        contactPhone:
          type: string
    PurchaseOrder:
      type: object
      required: [id, poNumber, supplierId, supplierName, items, totalAmount, status, orderDate, expectedDelivery]
      properties:
        id:
          type: integer
        poNumber:
          type: string
        supplierId:
          type: integer
        supplierName:
          type: string
        items:
          type: integer
        totalAmount:
          type: number
        status:
          type: string
        orderDate:
          type: string
        expectedDelivery:
          type: string
        actualDelivery:
          type: ["string", "null"]
        currency:
          type: string
    PurchaseOrderInput:
      type: object
      required: [supplierId, expectedDelivery, currency]
      properties:
        supplierId:
          type: integer
        expectedDelivery:
          type: string
        currency:
          type: string
        notes:
          type: string
    PurchaseOrderUpdate:
      type: object
      properties:
        status:
          type: string
        actualDelivery:
          type: string
...