# lib/api-zod/src/generated/types/inventoryItemUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Production Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface InventoryItemUpdate {
  description?: string;
  category?: string;
  warehouse?: string;
  onHandQty?: number;
  reorderPoint?: number;
  safetyStock?: number;
  unitCost?: number;
  status?: string;
}